Redesign interface to reflect new design
Overall the changes here better reflect the current design. Here's an overview of the core changes:

- modularize Python files supporting the interface
- use `unGhostAll` rather than unghosting a specific object
- use proxies instead of objects directly
- set start/end using visible timeline rather than min/max
- convert to independent euler curves before performing reduction
- rename `Select` button to `Evaluate`
- add combo box to choose OpenCL device
- implement stack based undo queue (see new `SavedAnimation` class)
